Conflicts:
	src/AppBundle/Mailer/Mailer.php

+++ src/AppBundle/Controller/ReservationController.php
@@ -24,13 +24,44 @@ class ReservationController extends Controller
      */
     public function indexAction()
     {
-        $em = $this->getDoctrine()->getManager();
+        if($this->container->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            $em = $this->getDoctrine()->getManager();
 
-        $reservations = $em->getRepository('AppBundle:Reservation')->findAll();
+            $reservations = $em->getRepository('AppBundle:Reservation')->findAll();
 
-        return $this->render('reservation/index.html.twig', array(
-            'reservations' => $reservations,
-        ));
+            return $this->render('reservation/index.html.twig', array(
+                'reservations' => $reservations,
+            ));
+        }else
+        {
+            $em = $this->getDoctrine()->getManager();
+
+            $reservations = $em->getRepository('AppBundle:Reservation')->findByUser($this->getUser());
+
+            return $this->render('reservation/index.html.twig', array(
+                'reservations' => $reservations,
+            ));
+        }
+
+    }
+
+    /**
+     * Lists all Reservation entities.
+     *
+     * @Route("/all", name="reservation_all")
+     * @Method("GET")
+     */
+    public function allAction()
+    {
+            $em = $this->getDoctrine()->getManager();
+
+            $reservations = $em->getRepository('AppBundle:Reservation')->findAll();
+
+            return $this->render('reservation/index.html.twig', array(
+                'reservations' => $reservations,
+            ));
+
+
     }
 
     /**
@@ -41,24 +72,26 @@ class ReservationController extends Controller
      */
     public function newAction(Request $request)
     {
-        $reservation = new Reservation();
-        $form = $this->createForm('AppBundle\Form\ReservationType', $reservation);
-        $form->handleRequest($request);
+            $reservation = new Reservation();
+            $form = $this->createForm('AppBundle\Form\ReservationType', $reservation);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($reservation);
-            $em->flush();
-            
-            $this->get('app.mailer')->sendMail($this->getUser()->getEmail());
-            
-            return $this->redirectToRoute('reservation_show', array('id' => $reservation->getId()));
-        }
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $reservation->setUser($this->getUser());
+                $em->persist($reservation);
+                $em->flush();
 
-        return $this->render('reservation/new.html.twig', array(
-            'reservation' => $reservation,
-            'form' => $form->createView(),
-        ));
+                $this->get('app.mailer')->sendMail($this->getUser()->getEmail());
+
+                return $this->redirectToRoute('reservation_show', array('id' => $reservation->getId()));
+            }
+
+            return $this->render('reservation/new.html.twig', array(
+                'reservation' => $reservation,
+                'form' => $form->createView(),
+            ));
+
     }
 
     /**
@@ -85,23 +118,27 @@ class ReservationController extends Controller
      */
     public function editAction(Request $request, Reservation $reservation)
     {
-        $deleteForm = $this->createDeleteForm($reservation);
-        $editForm = $this->createForm('AppBundle\Form\ReservationType', $reservation);
-        $editForm->handleRequest($request);
+        if($this->getUser() == $reservation->getUser() || $this->container->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            $deleteForm = $this->createDeleteForm($reservation);
+            $editForm = $this->createForm('AppBundle\Form\ReservationType', $reservation);
+            $editForm->handleRequest($request);
 
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($reservation);
-            $em->flush();
+            if ($editForm->isSubmitted() && $editForm->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($reservation);
+                $em->flush();
 
-            return $this->redirectToRoute('reservation_edit', array('id' => $reservation->getId()));
+                return $this->redirectToRoute('reservation_edit', array('id' => $reservation->getId()));
+            }
+
+            return $this->render('reservation/edit.html.twig', array(
+                'reservation' => $reservation,
+                'edit_form' => $editForm->createView(),
+                'delete_form' => $deleteForm->createView(),
+            ));
+        }else{
+            return $this->render('reservation/error.html.twig');
         }
-
-        return $this->render('reservation/edit.html.twig', array(
-            'reservation' => $reservation,
-            'edit_form' => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
-        ));
     }
 
     /**
@@ -112,16 +149,21 @@ class ReservationController extends Controller
      */
     public function deleteAction(Request $request, Reservation $reservation)
     {
-        $form = $this->createDeleteForm($reservation);
-        $form->handleRequest($request);
+        if($this->getUser() == $reservation->getUser() || $this->container->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            $form = $this->createDeleteForm($reservation);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($reservation);
-            $em->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($reservation);
+                $em->flush();
+            }
+
+            return $this->redirectToRoute('reservation_index');
+        }else{
+            return $this->render('reservation/error.html.twig');
         }
 
-        return $this->redirectToRoute('reservation_index');
     }
 
     /**
